Keyboard modularization

+++ src/message_handlers/erase_languages.py
@@ -1,7 +1,8 @@
 import telebot
 import fsm
+import bot_utils
 from flashcard import Word, Card
-from bot_utils import get_id, create_key_button, create_inline_key_button
+from bot_utils import get_id
 
 
 def handle_erase_languages(bot, rtd):
@@ -20,26 +21,13 @@ def handle_erase_languages(bot, rtd):
 			rtd.set_state(user_id, fsm.IDLE)
 			return 	
 		
-		btn = []
-		aux_counter = 0
-		for language in known_languages:
-			text = language
-			btn.append(create_inline_key_button(text, str(aux_counter)))
-			aux_counter += 1
-
-		markup = telebot.types.InlineKeyboardMarkup()
-
-		for i in range(0, len(btn)//2):
-			markup.row(btn[2*i], btn[2*i+1])
-
-		if len(btn) % 2 == 1:
-			markup.row(btn[len(btn)-1])
-
-		markup.row(create_inline_key_button("End selection", "DONE"))
+		btn = bot_utils.create_inline_keys_sequential(known_languages)
+		set_btn = set()
+		markup = bot_utils.create_selection_inline_keyboard(set_btn, btn, 3, ("End selection", "DONE"))
 
 		rtd.temp_user[user_id] = []
-		rtd.temp_user[user_id].append(set())
-		rtd.temp_user[user_id].append(known_languages)
+		rtd.temp_user[user_id].append(set_btn)
+		rtd.temp_user[user_id].append(btn)
 		bot.send_message(user_id, "Select languages to erase:", reply_markup=markup, parse_mode="Markdown")
 		rtd.set_state(user_id, fsm.next_state[fsm.IDLE]['erase_languages'])
 		
@@ -48,46 +36,21 @@ def handle_erase_languages(bot, rtd):
 							rtd.get_state(get_id(call.message)) == (fsm.ERASE_LANGUAGES, fsm.SELECT_LANGUAGES))
 	def callback_select_words(call):
 		user_id = get_id(call.message)
-		print("CALLBACK TEXT: {}   DATA: {}".format(call.message.text,call.data))
-		btn_number = call.data
-
+	
 		set_btn = rtd.temp_user[user_id][0]
-		languages = rtd.temp_user[user_id][1]
-		try:
-			btn_number = int(btn_number)
-			markup = telebot.types.InlineKeyboardMarkup()
-			btn = []	
-			if btn_number in set_btn:
-				set_btn.remove(btn_number)
-			else:
-				set_btn.add(btn_number)
-
-			for i in range(0, len(languages)):
-				if i in set_btn:
-					btn.append(create_inline_key_button(">>" + languages[i] + "<<", str(i)))
-				else:
-					btn.append(create_inline_key_button(languages[i], str(i)))
-
-			markup = telebot.types.InlineKeyboardMarkup()
-
-			for i in range(0, len(btn)//2):
-				markup.row(btn[2*i], btn[2*i+1])
-
-			if len(btn) % 2 == 1:
-				markup.row(btn[len(btn)-1])
-
-			markup.row(create_inline_key_button("End selection", "DONE"))
-			bot.edit_message_text(chat_id=user_id, message_id=call.message.message_id, text="Select languages to erase:", reply_markup=markup)
-			rtd.set_state(user_id, fsm.next_state[(fsm.ERASE_LANGUAGES, fsm.SELECT_LANGUAGES)]['continue'])
-
-		except ValueError:
-			#DONE
+		set_btn, done = bot_utils.parse_selection_inline_keyboard_ans(call.data, set_btn)
+		btn = rtd.temp_user[user_id][1]
+		
+		if done == True:
 			bot.delete_message(chat_id=user_id, message_id=call.message.message_id)
 			text = "_Erased languages:_\n"
 			for i in set_btn:
-				print(rtd.erase_language(user_id, languages[i]))
-				text += "*." + languages[i] + "*\n"
+				print(rtd.erase_language(user_id, btn[i][0]))
+				text += "*." + btn[i][0] + "*\n"
 			bot.send_message(user_id, text, parse_mode="Markdown")
 			rtd.set_state(user_id, fsm.next_state[(fsm.ERASE_LANGUAGES, fsm.SELECT_LANGUAGES)]['done'])
-
-
+		
+		else:
+			markup = bot_utils.create_selection_inline_keyboard(set_btn, btn, 3, ("End selection", "DONE"))
+			bot.edit_message_text(chat_id=user_id, message_id=call.message.message_id, text="Select languages to erase:", reply_markup=markup)
+			rtd.set_state(user_id, fsm.next_state[(fsm.ERASE_LANGUAGES, fsm.SELECT_LANGUAGES)]['continue'])
